DELTA windows-scache-deadlock-fix-20050827 AUTHOR jaltman@secure-endpoints.com
It was reported that Microsoft Word when editing files stored in AFS
would cause OAFW to fail to respond.  It was determined that a scp->mx
lock was not being released in buf_WaitIO if the no one was waiting
on the scp.

This patch corrects the deadlock and fixes some debugging messages.