DELTA windows-release-notes-20050925 AUTHOR jaltman@secure-endpoints.com
Update the installed documentation to include the new OpenAFS for
Windows release notes.   No IBM documentation is being removed but
the top level index page has been re-written to indicate that the IBM
docs are out of date.   The new OAFW Release Notes have been added to
the ReleaseNotes with a link.