DELTA STABLE14-windows-byte-range-locks-20051006 AUTHOR jaltman@secure-endpoints.com
Do not mark locks lost simply because the ExtendLock failed.
A lock is only lost if the server responds with EINVAL indicating
that the lock no longer exists. A lock can be renewed by other clients
that are also using the file.  The client can make no assumptions about
the status of a lock based upon the passage of time.