DELTA windows-drivemap-renewal-20051121 AUTHOR jaltman@secure-endpoints.com
When afscreds.exe performs a drive map renewal it should not delete
all existing drive mappings.  This undoes the efforts of the AFS
logon network provider and any persistent mappings created with the
explorer shell.   This patch extends the interface to provide a mode
in which DoMapShareChange does not remove unknown drive mappings.