DELTA viced-probe-uuid-on-old-connection-20060210 AUTHOR jaltman@secure-endpoints.com
When we have found an existing host via Uuid, send the probe to the
connection associated with the old host instead of the new one.  If
you send it to the new connection, we will fail to detect that the
old host is no longer responding and hence that it should be removed.
This should solve the 55 second delay problem seen from clients with
Uuids that move addresses or sit behind a NAT.