DELTA pretty-html-synopsis-20060228 AUTHOR rra@stanford.edu
Implement proper synopsis wrapping for HTML generation.

This was done in three pieces.  First, add HTML-specific tags to the POD to
mark the synopsis for HTML purposes so that we can apply style information
to it.  Second, update the style sheet to indent all lines except for the
first in the synopsis section.  Third, add the appropriate S<> tags around
option and argument pairs so that we don't wrap between the option and its
argument.

Unfortunately, due to the <I<foo>> style that looks nicer for other reasons,
we have to use the very verbose S<<< >>>.  Oh well.