DELTA sparc-linux-config-20060317 AUTHOR rra@stanford.edu
Collapse all of the sparc*_linux* cases into a single case.  Not just for
ease of maintenance -- this also results in clearer error messages for
configurations that aren't supported and should allow building of the
userspace tools for a few @sys types that the kernel build doesn't support.
Reported by Michael Conrad and Karl Ramm.  Patch has been in the Debian
package for some time.