DELTA STABLE14-xstat_fs-fix-20060407 AUTHOR jaltman@secure-endpoints.com
as pointed out by cg2v@andrew.cmu.edu the addition of performance
collection data for the GetCapabilities RPC broke backward compatibility
with previous builds because it altered the size of the data structure
sent on the wire.

Instead we count the number of GetCapabilties calls but do not gather
performance timing data as this can be done using spare fields.