DELTA STABLE14-windows-protect-against-vcp-undercount-20060408 AUTHOR jaltman@secure-endpoints.com
An undercount has been detected of the smb_vc_t objects stored
in the smb_allVCsp list.  Unfortunately, we have yet to be able
to find the cause of the undercount so this patch adds logic to
protect against the side effects until such time as the cause
can be identified.