DELTA windows-client-side-locking-20060420 AUTHOR asanka@secure-endpoints.com
In order to make the client side locking more usable, if the client
is denied a lock by the server and the user only has read privileges,
then we will allocate a local lock.  Local locks are not allocated
for write locks.

There is a bug in the file server locking dating back to at least
AFS 3.1 in which the lock privilege is used for read, write and insert
locks.  According to the docs, the lock privilege is only supposed to
control read locks.  Write locks on new and existing files are supposed
to be controlled by the Write and Insert privileges.  This will be fixed
in the file server by another commit.