DELTA DEVEL15-windows-no-freelance-vldb-failover-20060622 AUTHOR jaltman@secure-endpoints.com
When Freelance mode is not in use the cache manager must load the root.afs
volume of the default cell in order to be able to start.   Unfortunately,
if the first vlserver fails the request to obtain the volume information
will timeout and no retries will be performed.  This patch performs up
to 10 retries in the initialization code in order to enable the vlservers
to be marked down so that failover can occur.