DELTA windows-largefile-support-20060623 AUTHOR asanka@secure-endpoints.com
This patch adds large file support (> 2GB) to the CIFS server and
the cache manager.

It also corrects a significant oversight where the definition of
struct rx_call was inconsistent due to a failure to define AFS_PTHREAD_ENV
prior to including rx.h in all source files.  This prevented all errors
during rx_Read and rx_Write to be ignored.