DELTA vioc-header-split-20060802 AUTHOR rra@stanford.edu
Modify vice.h so that it doesn't depend on afs/param.h.  Split the VIOC*
constant definitions out of afs/venus.h into a separate header file
included by it.  Update various comments to reflect a more recent version
of reality.  This is in preparation for kopenafs work so that kopenafs.h
can include afs/vioc.h as well to get the same definitions.