DELTA windows-restrict-timeouts-to-cifs-20060808 AUTHOR jaltman@secure-endpoints.com
If afsd takes longer than the cifs timeout to complete an operation
the cifs client may break the virtual circuit and create a new one.
this forces all file handles to be closed and locks to be released.
Try to prevent the circuit from being destroyed by allowing the
CIFS lanmanworkstation SessTimeout value to enforce upper limits
on the Conn and Hard Dead Timeouts.   Permit this automatic configuration
to be disabled by setting specific configuration values for timeouts
in the registry.