DELTA DEVEL15-windows-delete-test-file-not-dir-20070131 AUTHOR jaltman@secure-endpoints.com
the cm_CheckNTDelete function was testing the permissions of the
directory and not the permissions of the file being deleted.
This resulted in inconsistencies between the attributes of the
file and the ability to delete it.

this patch also adds a large amount of trace log entries as there
was none in this section of the code before.