DELTA STABLE14-viced-write-perm-grants-read-lock-20070423 AUTHOR jaltman@secure-endpoints.com
the permission to obtain a write-lock is granted either by having the
"w" permission or the "i" when owning the file.  The permission to obtain
a read-lock has been granted by having the "k" permission.  However it
makes no sense that someone can obtain a write-lock but not a read-lock.
This patch grants permission to obtain read-locks to those who can obtain
write-locks.